Enable using BuildQueryPlanOptions in the gateway via the wasm package (#166)
This PR uses the updated API to thread through the `BuildQueryPlanOptions` object to wasm. It contains a single flag at the moment: `autoFragmentization`. 

This flag is now supported and implemented by the Rust query planner.

We also had to update the test snapshots to have the fragments sorted by name.