Fix crash when operation variables are not provided
Given a query like this:

query MyQuery($arg: Int) {
	myQuery(arg: $arg)
}

In case the arg variable is not provided on
the GraphQL request, even though it's not required,
the Gateway would crash, since it would try
to coearce a nullish variables object.
This behavior is not expected behavior,
since the GraphQL-JS library handle this
nicely [1].
In order to prevent further crashes, a check
was added it will simple provide an empty
object to the getVariablesValues(), just
like the original implementationl.

[1]: https://github.com/graphql/graphql-js/blob/d6e760cf7060217bd25fa934bd104ff2400aad96/src/execution/execute.js#L347