Lost directives on inline fragment (#178)
As reported in #177, we lose any directives on inline fragments from the original query

## Problem

We create new scopes either when we drill into object fields, inline fragments, or fragment spreads.

When we go into object fields, we collect the entire field node, so we don't lose directives.
We do however, lose directives when we create a new scope for inline fragments and fragment spreads.

## Solution for inline fragments

This PR fixes #177 by adding in any directives we find when collecting inline fragments and adding them to the scope we create for any fields in the fragment.

When we rebuild the operation that goes to downstream services, we wrap them in inline fragments if needed, if the fields we're wrapping have optional directives on the scope they're in, we add it to the inline fragment we're creating.

## Fragment spreads

As far as I can tell, we have the same problem with fragment spreads and directives on fragment definitions because the planner unfurls any fragments. I've opened #179 for this case.