Create stargate Dockerfile and GH Publish action
This commit introduces a Dockerfile for building a stargate image.
Currently this is building from an in-flight PR (#135), which we
can update as soon as it lands on `main`.

The additional GH action publishes the container to the new
GitHub Container Registry. For now, this requires a PAT (as noted
in the comments). We should remove this in favor of GITHUB_TOKEN
as soon as it's supported.