Use `monorepo-prepare` rather than `preinstall` script.
Follows-up apollographql/federation#184

This adjusts the life-cycle hook which is meant for building within the
monorepo to use the `monorepo-prepare` hook, which is only run when
installing this monorepo for local development, and is invoked by the
top-level `package.json`'s `postinstall` script in the root of the monorepo.

The other possible option here which actually makes the most sense is
`prepare`, but unfortunately, that breaks our CI pipeline right now since it
requires running it prior to the MOCK publishing that we do which generates
tarballs. See [[Failure]].

Fixes: https://github.com/apollographql/federation/issues/184

[Failure]: https://app.circleci.com/pipelines/github/apollographql/federation/219/workflows/3577e277-1e21-4d82-a76c-ad0e39b27064/jobs/549