Remove internal dependency on `schema` from composition
This commit cuts the gateway over to relying solely on
the `composedSdl` output from composition. This unlocks
the ability for either a user or Apollo Studio to serve
the gateway CSDL instead of each SDL file per service.

This is fairly straightforward, as `schema` was only
depended on in one place by query plan execution, during
the final pass of graphql-js `execute()`. Instead, we now
construct a `GraphQLSchema` ourselves from the CSDL. This
comes with a caveat: the schema we're creating doesn't have
all of the annotations that once existed in the schema that
we were using from `composeAndValidate()`. However, this is
inconsequential for the way the schema is currently being
used during the final execution pass.