fix(infra): Force publish wasm query planner for now (#298)
Admittedly, this will take us back to the dark ages of over-publishing from
whence we recently came (#194), but fortunately it is with a package that's
really only intended to be used internally for now.

One might recall a recent botched release (see: 0328553 004b8a1) in which
the wasm package didn't get the required version bump that it needed. This is
possibly due to a restructuring of the repository, but my suspicion is that
this runs a bit deeper and the workflow we're hoping for isn't achievable by
typical uses of lerna.

For additional context, but likely unnecessary reading:
To elaborate a bit, we need to map changes in the rust query planner into a
resulting lerna changed entry in the @apollo/query-planner-wasm package (which
is the artifact of changes to the rust query planner). This is proving
difficult and may require some digging and creativity.

All that is to say, at the cost of over-publishing for now, there shouldn't be
any more failed releases due to this issue.