Remove use of `moduleNameMapper` for local imports
Previously, we used Jest's `moduleNameMapper` option to import files from their `src` instead of their `dist` directory. That means we rely on `ts-jest` to compile not just the package currently under test, but also any imported code from other packages in the same repository.

The main benefit of that was that we didn't have to compile all packages before each test run. Jest would only compile what was imported, and only when the file had changed (because Jest includes its own file cache). That did sometimes lead to issues however, because the file cache doesn't correctly keep track of dependencies. `ts-jest` was also often unable to apply the right `tsconfig.json` to imported projects.

Nowadays, `tsc`'s support for incremental compilation means that whole project compiles are really fast. Relying on `tsc` to compile our code is often even faster than the per file compilation that `ts-jest` is forced to perform. And it means dependencies are tracked correctly and the right configs are applied. The project is compiled before each run of `npm test` through a `pretest` script, so the code your tests import should never be out of date.