Update tests and gateway stop() behavior
This is mostly a fairly large update / revamp to the networkRequests
test file. Previously skipped /flaky tests are now up-to-date and
functioning as expected!

Some smaller changes were made to the gateway: i.e. the stop()
function was updated to actually do something. It doesn't stop the
gateway process or even from serving requests, but it does stop it
from polling. This behavior was needed for the sake of testing so that
gateways still running in tests didn't overpoll and ruin network
mocks.

^additional commentary on above: this was only my suspicion and my
changes seem to have resolved the issue. Though I don't fully
understand why this wasn't a problem in the first place - or maybe
it was and was causing the flakiness in the past? I don't know.