Remove experimental, unsurfaced code coverage implementation.
Most of all, its worth noting that the code coverage results are not
surfaced _anywhere_ right now as of this commit and there are no thresholds
that hold us accountable for meeting or regressing on coverage requirements.
However, the motivating factor in this removal is an error in the CI because
of linker issues with profiling mechanisms that don't exist in WASM code.
Put simply, this is not able to provide coverage for the code that we trust
today.

We absolutely do WANT code coverage, but the link to the [Stackoverflow post]
in the comment within the code that introduced this current implementation
(which this commit removes) states a very real obstacle and a very relevant
issue.

The introduction of this was arguably a bit premature given the relative
infancy of the Rust code coverage tooling in the ecosystem and we never
agreed on how we wanted to configure this.  Hope is on the way (see
attached), but for now let's just count on the fact that Rust's compiler is
pretty good and Clippy (which we also have enabled on this repository)
already points out dead-code and omissions to best practices.
Alternatively, let's actually discuss how we want to configure this to
really surface code coverage problems in the code we have in a meaningful
way, not just running code coverage in CI for the sake of running code
coverage in CI but not giving any meaningful output that indicates if it was
coverage win or loss or anything.)

Ref: https://blog.rust-lang.org/inside-rust/2020/11/12/source-based-code-coverage.html
[Stackoverflow post]: https://stackoverflow.com/questions/61989746/how-do-i-use-zprofile-to-generate-code-coverage-for-a-nostd-kernel)