Convey missing prerequisites, but don't try to install them.
There are various prerequisites necessary for actually installing and
building this monorepo, including Node.js, a compiler toolchain with
make/cmake, Python, Rust, etc.

Node.js and Rust tooling (npm and cargo, respectively) take care of
installing packages that this project depends on but the burden of
installing most the more core prerequisites for this repository often
managed at a higher level and varies widely depending on operating system
and user preferences.

For example, the Bash shell script that this commit removes doesn't work on
Windows (invoking it fails, because Bash) and running the suggested
install command globally installs Rust for the user.

Additionally, there are various ways that users like to manage these
installations on their systems using existing package managers.  For example,
Rust can be installed globally using the Rust installer, or it can be
installed using Homebrew and contained to its mechanism for managing
installs.  Similarly, Node.js could be installed globally, but it can also
be managed by nvm.

Suggesting that the user need to install Rust is an excellent suggestion to
make, but the burden of providing installation instructions for various
platforms (and maintaining those instructions) seems like it should be that of
the Rust project.

Relates to https://github.com/apollographql/federation/pull/445