Move "ensure-prereqs" to the "preinstall" step for early failure.
If things are going to fail eventually because of a missing dependency,
failing quickly seems like a win.  By moving this to the preinstall, we
ensure that the user gets informed of missing dependencies quickly and can
act on resolving them.

Additionally, if any error what-so-ever happens in the install process, the
`postinstall` scripts won't be invoked.  That's a good thing for most cases,
but prerequisites for installation seem like they could be ensured in a
`preinstall` step.