fix(gateway): Ignore thrown errors from `extendSchema` (#478)
extendSchema throws errors in the case that we extend a type before it exists
during composition. This should certainly result in composition errors (which
we have validations for), but it shouldn't cause `composeAndValidate` to
throw, which is unexpected.

This change catches and discards errors thrown by `extendSchema`. The reason
for discarding and not capturing the error ("Unknown type Bar.") is that it's
duplicated when we capture errors from `validateSDL` immediately after.

In the `extendSchema` code path, there is only one throw case (excluding an
initial `assertSchema` which I'm not concerned about) and this test case
exercises it. However, I'm unsure about this approach because there could be
other errors thrown in the future which may be valuable or worth capturing.
This change may be worth revisiting in the case that `extendSchema` starts
throwing other errors which are useful in this context.