Update __tests__ tsconfig.json to include mocks
This `__tests__` directory specific `tsconfig.json` is only used within VS Code, to work around due to microsoft/vscode/#12463.

Unfortunately, that means we need to specify `__mocks__` as an explicit dependency because it now lives in a separate TypeScript project. (We use the package root `tsconfig.test.json` from Jest, and that includes all `__tests__` and `__mocks__` subdirectories in one project.)