Upgrade graphql-js to v15
This is a backport from #453. It appears that when `@apollo/federation` is
published without this and combined with `graphql@15` in a TypeScript setting,
compilation fails; with this upgrade, the produced `.d.ts` files are compatible
with both v14 and v15. See #537.

Doing this as a cherry-pick on a release branch to avoid no-op publishes of
`@apollo/gateway`.