limit to engines.node <15.x and engines.npm 6.x
This is now used to control the npm version that is used for installing
updates.  It also establishes the maximum intended supported (tested)
version of Node.js.

I (personally) fully support us making the migration to npm@7 at this point,
but it should be done in an structured, intentional manner and this just
makes the status-quo more clear to tooling that is just going to use the
latest version if it doesn't know better.

As for newer Node.js versions, that also seems fair, but we need to make
adjustments to a number of places to make that transition properly (e.g.,
tests, reviewing if we might also deprecate a version, etc.).