gateway: stop using apollo-server-testing
I've been making the assertion in various issues on `apollo-server` that I don't
feel like `apollo-server-testing` has much value over using the
`executeOperation` method (which was added to AS as part of implementing
`apollo-server-testing`) directly. I figured it was worth trying out my
suggestion over on this repo (esp when I noticed #553 upgrading
`apollo-server-testing`). My opinion is that the version without
`createTestClient` is more clear in these particular tests in that adding an
intermediate `call` object only serves to obfuscate.

The one nice thing that `createTestClient` does is allow you to pass an AST
rather than a string for the operation. We could change `executeOperation` to
support this directly too, though in this case just changing the tests to not
create AST nodes was easy enough.