feat(gateway): Update managed federation to fetch "cloud config" (#458)
This commit moves the gateway over to a simplified startup and updating
mechanism in managed mode. With the gateway now able to consume CSDL as input,
it's no longer necessary for the gateway to perform a series of fetches in
order to collect each service's partial SDL and do composition on the fly.

In order to take full advantage of this capability, we will serve CSDL straight
to the gateway rather than all of the individual partials.

This feature is being implemented as opt-in only for now, and should not be used
by the general public. The means to opt-in is by providing appropriate
configuration by either defining:
Via gateway configuration: experimental_schemaConfigDeliveryEndpoint: string OR
Via env var: APOLLO_SCHEMA_CONFIG_DELIVERY_ENDPOINT