renovate: remove `force` declaration (#587)
We already declare the same thing in the package.json `engines` section and apparently this force declaration is no longer needed (and can in fact be harmful, as it overrides the npm declaration for package.json). The same fix was applied to apollo-server by the creator of Renovate in https://github.com/apollographql/apollo-server/pull/5020 after we reported an issue.