Re-introduce last used TypeScript query planner code
This re-introduces the TypeScript query planner code as of https://github.com/apollographql/apollo-server/commit/60d974240769aac3e1c75ff968657b2d3e60e9b6, which is one of the parents of the commit that deleted that code from the `apollo-server` repository: https://github.com/apollographql/apollo-server/commit/149f78e683513485114d7076c7b183b650d883be.

I wanted to start out with this to then re-apply the changes I've made in an experimental repository in a more tractable manner.

On top of that, this commit adds a `composedSchema` directory, which contains code that is used to build a schema with Federation metadata from CSDL (this will be replaced with code that takes a core schema file with join directives as input). That was needed to get things wired up with the current gateway code.

 We temporarily export the same API we used for the wasm query planner from `@apollo/query-planner`, but implemented as a facade on top of the TypeScript one. This is ugly and inefficient (we shouldn't be parsing the query again), but the goal is to get things working first without making changes to the gateway code.