Print valid core schema and change `@http` to `@join__endpoint`
The generated core schema was invalid because the `@join__type` directive definition should be repeatable and have `key` as its argument (instead of `requires` and `provides`). We would also generate a `@join__field` directive without a graph name in case `serviceName` was undefined, which would lead to a syntax error when parsing.

This commit also renames `@http` to `@join__endpoint` and includes a directive definition for it (which was missing before).