Fix test that depends on the specific log message
AS 2.22 changes the log message a bit to be clear that it's a startup error (and
sometimes to suggest that you call start() yourself).

While I'm here, delete a process.exit mocking that does not seem to be
necessary, and validate the thrown error too.

Note that it might be reasonable to delete this test instead, or to delete the
validate of server.requestOptions.executor, rename the test, and move it
elsewhere?