Filter out null or undefined entities when executing fetch
If a parent fetch hasn't returned a value for an entity (because of a type condition for example), or has explicitly returned null, we want to filter out these entities before continuing with the fetch.

If there are no entities left we avoid the fetch altogether, but otherwise we'll perform the fetch with just the remaining entities.