Core schema support and Typescript query planner redux (#622)
This commit re-introduces the TypeScript query planner, and adds support for core schema as input. This builds on the work @trevor-scheer has done in #588 to have composition output core schema instead of the previous incarnation of CSDL.

We've verified the output of the query planner through a large number of tests and feel confident these changes shouldn't affect the generated query plans.

@apollo/query-planner currently exposes the same API as before, but we'll change this in a follow-up PR to avoid repeated printing/parsing of schemas and queries.

Any references to CSDL both internally and outward facing have been updated to their new counterpart:
`supergraphSdl`. Supergraph SDL implements both the core and join specs, necessary for powering a router.

Co-authored-by: Trevor Scheer <trevor.scheer@gmail.com>
Co-authored-by: Jesse Rosenberger <git@jro.cc>