chore(ci): Be explicit about cache restore keys
Currently, the `key`s are meant to be (and I think so) get hit first during
a restore, but it's not intuitive to believe that in reading this code.
That's really a GitHub Action DX thing in my opinion; their documentation does
suggest:

> The cache action first searches for cache hits for key and restore-keys in
the branch containing the workflow run.

Putting them explicitly should be more clear without additional cost though
(cache hits should be instant).