Change query planner API to avoid double parsing and schema building
We originally kept the same API we used for the wasm query planner, but that relied on passing in strings for the schema and operation. Since the query planner is now implemented in TypeScript again, we can avoid double parsing and schema building by passing in a composed schema and a parsed operation.