buildQueryPlan: FragmentMap can map to undefined
The previous type implied that literally any fragment name can be used to retrieve a fragment from a fragment map. Typically, GraphQL documents contain a finite number of fragments.

The only code that reads from it happened to check the value for truthiness, but this wasn't enforced by the type system.