buildQueryPlan: clarify typing on two more "map"-type objects
Like in #633, it doesn't help our type-checking to tell TypeScript that
literally every variable name in a VariableUsages or variableDefinitions maps to
a defined output.

For variableDefinitions, I'm fixing by keeping it as an Object and providing a
better error message in the (hopefully shouldn't-happen due to GraphQL
validation, but hey, we do allow Apollo Server users to override the validation
rule list) case that a variable usage has no corresponding definition.

For VariableUsages, that typing would mean that `Object.values` would return a
list whose TypeScript type could contain `undefined`, which it won't, so it's
better to switch over to a Map, since Map typing naturally differentiates
between the lack of a value and an explicit undefined.

Like #633, this probably doesn't fix any real bugs (unless we are given an
invalid operation to execute).