Remove Rust projects not under active development
We are currently moving forward with the original TypeScript query planner rather than
the late-2020 Rust query planner. We also have not actively worked on `stargate`
recently, and `graphql-parser` is only used by those packages. So for the time
being, remove them from git to simplify the repository and speed up CI.  (If we
find these project valuable later we can always restore them from history.)

This is NOT a complete removal of Rust from this repository: we still have
`harmonizer` and `harmonizer-bin` (which package `@apollo/federation` for use by
Rust code such as Rover). However, they consist mostly of wrapper code rather
than deep logic, and Rust does not need to be executed in order to do Node-only
work any more.

This consists of:
- Deleting the root directories for `stargate`, `query-planner`,
  `graphql-parser`, and `query-planner-wasm`
- Not setting up Rust for the CircleCI Node testing, because it's no longer
  needed to do Node stuff. This reduces the time to run CircleCI jobs from ~4
  minutes to ~1 minute.
- Removed `stargate`-only GH publish action.
- Remove the duplication of rust-slow-test.yml and rust-test.yml, leaving only
  rust-test.yml which is similar to the old rust-slow-test.yml (but without
  wasm-pack)
- Remove the requirement that rustup be installed in order to run npm things
- Remove grcov.sh (coverage seems less interesting for harmonizer than for real
  tests)

We should seriously consider the suggestion in #602 that harmonizer be moved to
a separate repository to allow this one to have no (slow) CI dependencies on
Rust.