Adds debug logging to the query planner code
To work on the query planner, it is useful to be able to follow the
steps taken when building a plan (and while a debugger can help, it can
be messy unless you already roughly know what you want to look at).

This commit introduce simple debug logging facilities and adds some
initial logging to the query planning code. More logging can easily be
added over time, or be manually added to narrow down a particular issue.

The added logging is not meant for end users, but only to help
debugging/improve the query planning code.

The debug logging is enabled by setting the `QP_DEBUG` environment
variable to "true". Once enabled, logs are printed on the console.

The introduced debug logging facility is a thin layer over `console.log`
that adds:
- conditional logging. Logging is only enabled if the `QP_DEBUG`
  variables is set.
- a simple way to indent messages (not unlike `console.group` but
  at tad more flexible). As the query planner algorithm works with
  a kind of depth-first recursion, indentation helps greatly improves
  legibility.
- lazy computation of the logged message through closures.

This commit adds a dependency on [chalk](https://www.npmjs.com/package/chalk)
for coloring the output.