Install a specific version of npm in the Rust builds
Currently, we're just using the version of npm that comes with Node.js
versions in order to avoid what I found to be a pretty annoying case on
Windows where npm couldn't be upgraded because it couldn't replace itself
because it was in use.

It seems to be common in the community to just work around this by: Only
relying on the version of npm that comes with Node.js or, alternatively, by
moving npm out of the way to make room for the new npm.

So I had an idea to use yarn, which is actually the same way npm had to be
upgraded to 6 for a while because of a bug in npm 5 that shipped with common
Node.js versions.

That idea didn't work.  Neither did another suggestion to use a utility that
claims to try tackling this.  I guess this is a Hard problem.

So I just do the most commonly accepted answer and move them out of the way,
use the _moved_ version to install a fresh version, and let the new one
appear in the place of the original.  No cleanup because we're in CI.