Assert Node.js versions and npm versions in Rust tests
Previously, we just printed out the versions.  That's useful, but in my
testing, I found it useful to just make sure that the logic we're using
worked reliably.

Now we make sure they match exactly.

While this gives us another version to update `NODE_VERSION` and
`NPM_VERSION`, that's something that Renovate can take care of for us (not
sure immediately if it will just do it without a config change though).