Typecheck tests before running them (#727)
Something like this was intended to be part of #453. Because we have
`diagnostics: false` in jest.config.base.js, typechecking failures in
ts-jest don't actually stop the tests from running successfully; this
change makes us typecheck the tests before we run them. (If you really
want to run tests that don't typecheck you can run `npx jest` directly.)

As part of this, fix a typechecking failure caused by the upgrade in
PR #716 and run `npm dedup` to fix another typechecking failure. The `npm dedup`
itself causes a snapshot to need to be updated (as the snapshot was created by
some code that was still running a nested apollo-server 2.23).

We may want to consider finding a way to be comfortable removing
`diagnostics: false` instead as in #661, but we'll need to fix #662 before
we can do that.