chore(infra): Upgrade monorepo to use npm@7 and enable `engine-strict` (#708)
This upgrades us to use `npm@7` at both the root and in `docs/` and enables the
`engine-strict` option via a root-level `.npmrc` file in this repository.

The intent of introducing `engine-strict` is to make sure that the users of
this monorepo are not led astray into using an unsupported version of Node.js
or npm that might cause unexpected side-effects. (Running the wrong version
of `npm` is the most motivating reason today!)

Users who wish to run this monorepo with a version of Node.js and npm that are
not specified in the `engines` range may still do so by running their commands
with the `--no-engine-strict` flag to acknowledge that there might be unexpected
or untested results/outcomes from their actions.  e.g, `npm install --no-engine-strict`.